# QSDM/docs/docs/openapi.yaml
openapi: 3.0.3
info:
  title: QSD HTTP API
  description: |
    Quantum-Secure Dynamic Mesh Ledger (QSD, formerly QSD) HTTP API with
    military-grade security. Native coin: Cell (CELL), 8 decimals, smallest
    unit `dust`.
    
    All authenticated endpoints require:
    - JWT Bearer token in Authorization header
    - Request signing for POST/PUT/DELETE requests (X-Timestamp, X-Nonce, X-Signature headers)
    
    Security Features:
    - TLS 1.3 encryption
    - Quantum-safe JWT tokens (CRYSTALS-Dilithium)
    - Request signing with quantum-safe signatures
    - Rate limiting: default **100 requests per client per minute** (`[api] rate_limit_max_requests` / `rate_limit_window` or `QSD_API_RATE_LIMIT_MAX` / `QSD_API_RATE_LIMIT_WINDOW`; the legacy `QSD_*` env vars continue to be accepted during the deprecation window, see `REBRAND_NOTES.md`). Some routes are stricter (e.g. GET /monitoring/ngc-challenge at 15/min). **GET /health** and **/health/*** are exempt so probes are not throttled.
    - Replay attack protection
    - Role-based access control
    
    **Submesh policy (optional):** When the node loads `[network] submesh_config` / `QSD_SUBMESH_CONFIG` (legacy alias: `QSD_SUBMESH_CONFIG`), some ledger routes may return **422 Unprocessable Entity** if fee+geotag do not match a profile or the serialized payload exceeds `max_tx_size` (wallet send), or mint/token-create exceeds the strictest profile cap. See main README “Submesh profiles”.

...
  version: 1.1.0
  contact:
...
servers:
...
paths:
...
  /governance/params:
    get:
      tags:
        - Governance
      summary: Governance parameters snapshot
      description: |
        Read companion to the off-chain `QSD/gov/v1` write path that
        flows through `/transactions`. Returns 503 until a provider
        is installed via `api.SetGovernanceProvider(...)`. See
        `MINING_PROTOCOL_V2.md` §9.4.
      operationId: governanceParams
      responses:
        '200':
          description: Governance parameter snapshot
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '503':
          description: Governance provider not wired
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token signed with CRYSTALS-Dilithium (quantum-safe).
        Token format: base64(claims) + "." + base64(signature)
        
        For POST/PUT/DELETE requests, also include:
        - X-Timestamp: Unix timestamp (seconds)
        - X-Nonce: Unique nonce (32 bytes, base64)
        - X-Signature: Request signature (base64, quantum-safe)
...